requirements_all.txt
Package                Version
---------------------- --------
asgiref                3.6.0
certifi                2023.5.7
cffi                   1.15.1
charset-normalizer     3.1.0
coreapi                2.3.3
coreschema             0.0.4
cryptography           41.0.1
defusedxml             0.7.1
Deprecated             1.2.14
Django                 4.2.1
django-filter          23.2
django-oauth-toolkit   2.3.0
django-rest-swagger    2.2.0
djangorestframework    3.14.0
drf-social-oauth2      2.1.3
ecdsa                  0.18.0
idna                   3.4
itypes                 1.2.0
Jinja2                 3.1.2
jwcrypto               1.5.0
MarkupSafe             2.1.3
oauth2-provider        0.0
oauthlib               3.2.2
openapi-codec          1.3.2
pip                    23.1.2
psycopg2               2.9.6
pyasn1                 0.5.0
pycparser              2.21
PyJWT                  2.7.0
python-jose            3.3.0
python3-openid         3.2.0
pytz                   2023.3
requests               2.31.0
requests-oauthlib      1.3.1
rsa                    4.9
setuptools             65.5.1
simplejson             3.19.1
six                    1.16.0
social-auth-app-django 5.2.0
social-auth-core       4.4.2
sqlparse               0.4.3
tzdata                 2023.3
uritemplate            4.1.1
urllib3                2.0.3
wheel                  0.38.4
wrapt                  1.15.0
